feat : flyway 적용

+++ src/main/java/com/walgwalg/backend/core/service/WalkServiceInterface.java
@@ -14,7 +14,7 @@ public interface WalkServiceInterface {
     Map<String, String> startWalk(String userid, Date walkDate,String location, String address);
     void addGps(String userid, String walkId, String latitude, String longitude);
     List<ResponseGps.gps> getGps(String walkId);
-    void registerWalk(String userid,MultipartFile course,String walkId, Integer step_count,Integer distance, Integer calorie, String walkTime) throws ParseException;
+    void registerWalk(String userid,MultipartFile course,String walkId, Long step_count,Long distance, Long calorie, String walkTime) throws ParseException;
     void deleteWalk(String userId, String walkId);
     List<ResponseWalk.list> getAllMyWalk(String userid);
     Map<Date, ResponseWalk.calendar> getWalkForCalendar(String userid);
